user api,use bcrypt&jsonwebtoken

+++ server/interface/article.js
@@ -1,11 +1,13 @@
 import mongoose from 'mongoose'
 import Router from 'koa-router'
 import Article from '../db/models/article'
+import config from '../../config'
+import {verifyToken} from '../utils'
 
 const ObjectId = mongoose.Types.ObjectId
 
 const router = new Router({
-  prefix: '/article'
+  prefix: config.basePrefix + '/article'
 })
 
 // 获取文章列表
@@ -41,7 +43,6 @@ router.get('/getArticleList', async ctx => {
   ])
   if (totalResult.length) {
     total = totalResult[0].total
-
   }
 
   const article = await Article.aggregate([
@@ -50,7 +51,7 @@ router.get('/getArticleList', async ctx => {
     },
     {
       $lookup: {
-        from: 'categories',
+        from: 'category',
         localField: 'category_id',
         foreignField: '_id',
         as: 'category'
@@ -86,8 +87,8 @@ router.get('/getArticleList', async ctx => {
   ])
   if (article.length) {
     ctx.body = {
-      code: 0,
-      message: 'success',
+      code: config.SUCCESS_CODE,
+      message: '文章列表获取成功',
       count: article.length,
       total,
       data: article
@@ -95,7 +96,7 @@ router.get('/getArticleList', async ctx => {
     return
   }
   ctx.body = {
-    code: -1,
+    code: config.INFO_ERROR_CODE,
     message: '文章列表为空',
     count: 0,
     total: 0
@@ -113,7 +114,7 @@ router.get('/getDetail', async ctx => {
         },
         {
           $lookup: {
-            from: 'categories',
+            from: 'category',
             localField: 'category_id',
             foreignField: '_id',
             as: 'category'
@@ -140,25 +141,25 @@ router.get('/getDetail', async ctx => {
       ])
       if (detail.length) {
         ctx.body = {
-          code: 0,
+          code: config.SUCCESS_CODE,
           message: '文章详情获取成功',
           data: detail
         }
         return
       }
       ctx.body = {
-        code: -1,
+        code: config.INFO_ERROR_CODE,
         message: '文章详情获取失败，ID不正确'
       }
     } catch (e) {
       ctx.body = {
-        code: 1,
+        code: config.SYSTEM_ERROR_CODE,
         message: e.message
       }
     }
   } else {
     ctx.body = {
-      code: -1,
+      code: config.INFO_ERROR_CODE,
       message: '文章ID为空'
     }
   }
@@ -170,7 +171,7 @@ router.get('/getRandomArticle', async ctx => {
   limit = parseInt(limit)
   if (typeof limit !== 'number') {
     ctx.body = {
-      code: -1,
+      code: config.INFO_ERROR_CODE,
       message: 'limit参数必须是数字'
     }
     return
@@ -195,7 +196,7 @@ router.get('/getRandomArticle', async ctx => {
     ])
     if (list.length) {
       ctx.body = {
-        code: 0,
+        code: config.SUCCESS_CODE,
         message: '获取随机文章成功',
         count: list.length,
         data: list
@@ -203,13 +204,13 @@ router.get('/getRandomArticle', async ctx => {
       return
     }
     ctx.body = {
-      code: 0,
+      code: config.INFO_ERROR_CODE,
       message: '随机文章为空',
       count: 0
     }
   } catch (e) {
     ctx.body = {
-      code: 1,
+      code: config.SYSTEM_ERROR_CODE,
       message: e.message
     }
   }
@@ -217,11 +218,14 @@ router.get('/getRandomArticle', async ctx => {
 
 // 添加文章
 router.post('/addArticle', async ctx => {
+  let verifyResult = verifyToken(ctx)
+  if (!verifyResult) return
+
   const {title, content, category_id, author, thumbnail, description, keyword, importance = 1} = ctx.request.body
 
   if (!category_id) {
     ctx.body = {
-      code: -1,
+      code: config.INFO_ERROR_CODE,
       message: '分类为空'
     }
     return
@@ -229,14 +233,14 @@ router.post('/addArticle', async ctx => {
 
   if (!title) {
     ctx.body = {
-      code: -1,
+      code: config.INFO_ERROR_CODE,
       message: '标题为空'
     }
     return
   } else {
     if (await Article.findOne({title})) {
       ctx.body = {
-        code: -1,
+        code: config.INFO_ERROR_CODE,
         message: '标题已存在'
       }
       return
@@ -246,12 +250,12 @@ router.post('/addArticle', async ctx => {
   try {
     await new Article({title, content, category_id, author, thumbnail, description, keyword, importance}).save()
     ctx.body = {
-      code: 0,
+      code: config.SUCCESS_CODE,
       message: '文章添加成功'
     }
   } catch (e) {
     ctx.body = {
-      code: 1,
+      code: config.SYSTEM_ERROR_CODE,
       message: e.message
     }
   }
@@ -260,39 +264,51 @@ router.post('/addArticle', async ctx => {
 // 修改文章
 router.post('/updateArticle', async ctx => {
   const articleData = ctx.request.body
-  const result = await Article.updateOne({_id: articleData._id}, articleData)
-  if (result.ok) {
-    ctx.body = {
-      code: 0,
-      message: '修改文章成功'
+  let verifyResult = verifyToken(ctx)
+  if (!verifyResult) return
+  const _id = articleData.id
+  if (_id) {
+    const result = await Article.updateOne({_id}, articleData)
+    if (result.ok) {
+      ctx.body = {
+        code: config.SUCCESS_CODE,
+        message: '修改文章成功'
+      }
+      return
     }
-    return
-  }
-  ctx.body = {
-    code: 1,
-    message: '修改文章失败'
+    ctx.body = {
+      code: config.SYSTEM_ERROR_CODE,
+      message: '修改文章失败'
+    }
+  } else {
+    ctx.body = {
+      code: config.INFO_ERROR_CODE,
+      message: '文章ID为空'
+    }
   }
 })
 
 // 删除文章
 router.post('/deleteArticle', async ctx => {
+  let verifyResult = verifyToken(ctx)
+  if (!verifyResult) return
   const {id} = ctx.request.body
   if (id) {
     const result = await Article.deleteOne({_id: id})
     if (result.ok) {
       ctx.body = {
-        code: 0,
+        code: config.SUCCESS_CODE,
         message: '删除文章成功'
       }
     } else {
       ctx.body = {
-        code: 1,
+        code: config.SYSTEM_ERROR_CODE,
         message: '删除文章失败'
       }
     }
   } else {
     ctx.body = {
-      code: -1,
+      code: config.INFO_ERROR_CODE,
       message: '文章ID为空'
     }
   }
@@ -300,23 +316,25 @@ router.post('/deleteArticle', async ctx => {
 
 //  批量删除文章
 router.post('/deleteArticleList', async ctx => {
+  let verifyResult = verifyToken(ctx)
+  if (!verifyResult) return
   const {list} = ctx.request.body
   if (list.length) {
     const result = await Article.deleteMany({_id: {$in: list}})
     if (result.ok) {
       ctx.body = {
-        code: 0,
+        code: config.SUCCESS_CODE,
         message: '批量删除文章成功'
       }
     } else {
       ctx.body = {
-        code: 1,
+        code: config.SYSTEM_ERROR_CODE,
         message: '批量删除文章失败'
       }
     }
   } else {
     ctx.body = {
-      code: -1,
+      code: config.INFO_ERROR_CODE,
       message: '提交删除列表为空'
     }
   }
